refactor: move creation of tab menu to CreateTabMenu.js
