+++ src/index.js
@@ -1,5 +1,5 @@
 import createProductManage from './ProductManage/ProductManage.js';
-import createTop from './Top.js';
+import createTop from './Top/Top.js';
 
 createTop();
 createProductManage();
